Handle invalid IDs on related tags
Previously, the management command trusted that the IDs in the related tags existed, but this is not the case if there are some broken tags.

This leaves the related pages in a broken state, so we rather make sure that the ID in the tag exists, before adding it to the related pages.