Remove cache from list endpoints.
This cache was causing issues with the list endpoints being out of date,
since it was set with a very long TTL.

There's no way to invalidate this cache on changes, since the key is
derived from query string parameters, and Accept and Cookie headers in the
request.

If it's needed in the future, we will add it back, either with a very
short timeout, or a more direct cache that will cache only the things we
need, with cache keys that we can delete on changes.