Add empty value on content import form
For the content import, if the importer receives a "None" value for the
locale, then it imports all locales present in the import file.

This allows the locale field in the form to be empty, and adds a label
for it that makes it clear to the user that leaving the dropdown in the
"empty" position will result in the importer importing all locales
present in the file.